requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# Document Processing
paddleocr==2.7.3
paddlepaddle==2.6.0
pdf2image==1.17.0
pypdf==4.0.1
Pillow==10.2.0
opencv-python==4.9.0.80

# Layout Detection
layoutparser==0.3.4
torch==2.1.2
torchvision==0.16.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# LLM & RAG
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.3.1

# VLM for tables/charts
transformers==4.37.2
qwen-vl-utils==0.0.2

# Utilities
numpy==1.26.3
pandas==2.2.0
pydantic==2.6.0
tqdm==4.66.1
